requirements.txt
openai==1.77.0
python-dotenv==1.1.0
pydantic==2.11.3
